refactor: organize PDF link extraction functions in `fast` strategy
